Switch tests to use `@BeforeAll` and `@AfterAll` annotations
This will run the setup and tear down functions only once per test class, instead of between each test method. Reducing the amount of times test values need to be inserted in the DB for integration tests. This however, revealed that the model equality methods were doing comparisons to memory addresses and not comparing the state between the two compared models, so that is also addressed here.